Added/moved all commands as defined in google doc.
