+++ test/commands/build.test.js
@@ -1,0 +1,17 @@
+const {expect, test} = require('@oclif/test')
+
+describe('build', () => {
+  test
+  .stdout()
+  .command(['build'])
+  .it('runs hello', ctx => {
+    expect(ctx.stdout).to.contain('hello world')
+  })
+
+  test
+  .stdout()
+  .command(['build', '--name', 'jeff'])
+  .it('runs hello --name jeff', ctx => {
+    expect(ctx.stdout).to.contain('hello jeff')
+  })
+})
