+++ test/commands/prune.test.js
@@ -1,0 +1,17 @@
+const {expect, test} = require('@oclif/test')
+
+describe('prune', () => {
+  test
+  .stdout()
+  .command(['prune'])
+  .it('runs hello', ctx => {
+    expect(ctx.stdout).to.contain('hello world')
+  })
+
+  test
+  .stdout()
+  .command(['prune', '--name', 'jeff'])
+  .it('runs hello --name jeff', ctx => {
+    expect(ctx.stdout).to.contain('hello jeff')
+  })
+})
